Rename default_config to config
